refactor: extract form to the external component
